+++ RodTracker/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [v0.6.6]
 
 ### Added
-- a couple of warnings added which were missing in the last release
+- a couple of warnings (when data is loaded only from one camera) added which were missing in the last release
+- added a warning when no particles are detected in a given frame 
 - new flake8 version warnings (F824) suppressed (will be addressed in the new releases)
 
 ### Fixed